+++ app/Extensions/Gallery/lang/fr/gallery.php
@@ -2,19 +2,28 @@
 
 declare(strict_types=1);
 
-return  [
-  'object' =>
-   [
-    'photo' =>
-     [
-      'title' =>
-       [
-        'label' => 'Titre Updated',
-      ],
-    ],
-  ],
-  'action' =>
-   [
+return array (
+  'action' => 
+  array (
     'upload' => 'Envoyer une photo',
-  ],
-];
+  ),
+  'msg' => 
+  array (
+    'confirm_delete_photo' => 'Suppression de la photo ?',
+  ),
+  'object' => 
+  array (
+    'photo' => 
+    array (
+      'title' => 
+      array (
+        'hint' => 'Titre de la photo',
+        'label' => 'Titre Updated',
+      ),
+    ),
+  ),
+  'title' => 
+  array (
+    'gallery_list' => 'Galerie',
+  ),
+);